requirements.txt
Django==4.2.5
djangorestframework==3.14.0
requests==2.31.0
nose==1.3.7
django-nose==1.4.7
mock==5.1.0
parameterized==0.9.0
